Move PrettifyNames tests to a new file
